+++ src/Icecave/Archer/Console/Command/Travis/BuildCommand.php
@@ -2,9 +2,9 @@
 namespace Icecave\Archer\Console\Command\Travis;
 
 use Icecave\Archer\Coveralls\CoverallsClient;
-use Icecave\Archer\Coveralls\CoverallsConfigManager;
-use Icecave\Archer\Support\Isolator;
+use Icecave\Archer\FileSystem\FileSystem;
 use Icecave\Archer\GitHub\GitHubClient;
+use Icecave\Archer\Support\Isolator;
 use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
@@ -16,22 +16,24 @@ class BuildCommand extends AbstractTravisCommand
     public function __construct(
         GitHubClient $githubClient = null,
         CoverallsClient $coverallsClient = null,
-        CoverallsConfigManager $coverallsConfigManager = null,
+        FileSystem $fileSystem = null,
         Isolator $isolator = null
     ) {
         if (null === $githubClient) {
             $githubClient = new GitHubClient;
         }
+
         if (null === $coverallsClient) {
             $coverallsClient = new CoverallsClient;
         }
-        if (null === $coverallsConfigManager) {
-            $coverallsConfigManager = new CoverallsConfigManager;
+
+        if (null === $fileSystem) {
+            $fileSystem = new FileSystem;
         }
 
         $this->githubClient = $githubClient;
         $this->coverallsClient = $coverallsClient;
-        $this->coverallsConfigManager = $coverallsConfigManager;
+        $this->fileSystem = $fileSystem;
 
         parent::__construct($isolator);
     }
@@ -50,14 +52,6 @@ class BuildCommand extends AbstractTravisCommand
     public function coverallsClient()
     {
         return $this->coverallsClient;
-    }
-
-    /**
-     * @return CoverallsConfigManager
-     */
-    public function coverallsConfigManager()
-    {
-        return $this->coverallsConfigManager;
     }
 
     /**
@@ -132,11 +126,12 @@ class BuildCommand extends AbstractTravisCommand
 
         $coverallsExitCode = 0;
         if ($publishCoveralls) {
-            $coverallsConfigPath = $this
-                ->coverallsConfigManager()
-                ->createConfig($archerRoot, $packageRoot);
-
             $output->write('Publishing Coveralls data... ');
+            $coverallsConfigPath = $packageRoot . '/.coveralls.yml';
+
+            if (!$this->isolator->file_exists($coverallsConfigPath)) {
+                $this->isolator->copy($archerRoot . '/res/coveralls/coveralls.yml', $coverallsConfigPath);
+            }
 
             $coverallsExitCode = 255;
             $this->isolator->passthru(
@@ -167,15 +162,18 @@ class BuildCommand extends AbstractTravisCommand
             $command .= ' publish %s';
             $command .= ' %s/artifacts:artifacts';
             $command .= ' --message "Publishing artifacts from build %d."';
-            $command .= ' --coverage-image artifacts/images/coverage.png';
-            $command .= ' --coverage-phpunit artifacts/tests/coverage/coverage.txt';
-            $command .= ' --build-status-image artifacts/images/build-status.png';
-            $command .= ' --build-status-tap artifacts/tests/report.tap';
             $command .= ' --auth-token-env ARCHER_TOKEN';
-            $command .= ' --image-theme travis/variable-width';
-            $command .= ' --image-theme icecave/regular';
             $command .= ' --no-interaction';
             $command .= ' --verbose';
+
+            if ($publishCoveralls) {
+                // Remove test artifacts if coveralls is being used ...
+                $this->fileSystem->delete($packageRoot . '/artifacts/tests');
+            } else {
+                $command .= ' --coverage-image artifacts/images/coverage.png';
+                $command .= ' --coverage-phpunit artifacts/tests/coverage/coverage.txt';
+                $command .= ' --image-theme buckler/buckler';
+            }
 
             $command = sprintf(
                 $command,
@@ -203,5 +201,5 @@ class BuildCommand extends AbstractTravisCommand
 
     private $githubClient;
     private $coverallsClient;
-    private $coverallsConfigManager;
+    private $fileSystem;
 }