+++ test/suite/Icecave/Archer/Console/Command/Travis/BuildCommandTest.php
@@ -15,7 +15,7 @@ class BuildCommandTest extends PHPUnit_Framework_TestCase
 
         $this->githubClient = Phake::mock('Icecave\Archer\GitHub\GitHubClient');
         $this->coverallsClient = Phake::mock('Icecave\Archer\Coveralls\CoverallsClient');
-        $this->coverallsConfigManager = Phake::mock('Icecave\Archer\Coveralls\CoverallsConfigManager');
+        $this->fileSystem = Phake::mock('Icecave\Archer\FileSystem\FileSystem');
         $this->isolator = Phake::mock('Icecave\Archer\Support\Isolator');
 
         $this->application = new Application('/path/to/archer');
@@ -23,7 +23,7 @@ class BuildCommandTest extends PHPUnit_Framework_TestCase
         $this->command = new BuildCommand(
             $this->githubClient,
             $this->coverallsClient,
-            $this->coverallsConfigManager,
+            $this->fileSystem,
             $this->isolator
         );
 
@@ -61,7 +61,6 @@ class BuildCommandTest extends PHPUnit_Framework_TestCase
     {
         $this->assertSame($this->githubClient, $this->command->githubClient());
         $this->assertSame($this->coverallsClient, $this->command->coverallsClient());
-        $this->assertSame($this->coverallsConfigManager, $this->command->coverallsConfigManager());
     }
 
     public function testConstructorDefaults()
@@ -72,13 +71,10 @@ class BuildCommandTest extends PHPUnit_Framework_TestCase
             'Icecave\Archer\GitHub\GitHubClient',
             $this->command->githubClient()
         );
+
         $this->assertInstanceOf(
             'Icecave\Archer\Coveralls\CoverallsClient',
             $this->command->coverallsClient()
-        );
-        $this->assertInstanceOf(
-            'Icecave\Archer\Coveralls\CoverallsConfigManager',
-            $this->command->coverallsConfigManager()
         );
     }
 
@@ -166,15 +162,12 @@ class BuildCommandTest extends PHPUnit_Framework_TestCase
         $expectedWoodhouseCommand  = "/path/to/archer/bin/woodhouse publish 'Vendor/package'";
         $expectedWoodhouseCommand .= ' /path/to/project/artifacts:artifacts';
         $expectedWoodhouseCommand .= ' --message "Publishing artifacts from build 543."';
-        $expectedWoodhouseCommand .= ' --coverage-image artifacts/images/coverage.png';
-        $expectedWoodhouseCommand .= ' --coverage-phpunit artifacts/tests/coverage/coverage.txt';
-        $expectedWoodhouseCommand .= ' --build-status-image artifacts/images/build-status.png';
-        $expectedWoodhouseCommand .= ' --build-status-tap artifacts/tests/report.tap';
         $expectedWoodhouseCommand .= ' --auth-token-env ARCHER_TOKEN';
-        $expectedWoodhouseCommand .= ' --image-theme travis/variable-width';
-        $expectedWoodhouseCommand .= ' --image-theme icecave/regular';
         $expectedWoodhouseCommand .= ' --no-interaction';
         $expectedWoodhouseCommand .= ' --verbose';
+        $expectedWoodhouseCommand .= ' --coverage-image artifacts/images/coverage.png';
+        $expectedWoodhouseCommand .= ' --coverage-phpunit artifacts/tests/coverage/coverage.txt';
+        $expectedWoodhouseCommand .= ' --image-theme buckler/buckler';
 
         Phake::when($this->coverallsClient)
             ->exists('Vendor', 'package')
@@ -229,15 +222,12 @@ class BuildCommandTest extends PHPUnit_Framework_TestCase
         $expectedWoodhouseCommand  = "/path/to/archer/bin/woodhouse publish 'Vendor/package'";
         $expectedWoodhouseCommand .= ' /path/to/project/artifacts:artifacts';
         $expectedWoodhouseCommand .= ' --message "Publishing artifacts from build 543."';
-        $expectedWoodhouseCommand .= ' --coverage-image artifacts/images/coverage.png';
-        $expectedWoodhouseCommand .= ' --coverage-phpunit artifacts/tests/coverage/coverage.txt';
-        $expectedWoodhouseCommand .= ' --build-status-image artifacts/images/build-status.png';
-        $expectedWoodhouseCommand .= ' --build-status-tap artifacts/tests/report.tap';
         $expectedWoodhouseCommand .= ' --auth-token-env ARCHER_TOKEN';
-        $expectedWoodhouseCommand .= ' --image-theme travis/variable-width';
-        $expectedWoodhouseCommand .= ' --image-theme icecave/regular';
         $expectedWoodhouseCommand .= ' --no-interaction';
         $expectedWoodhouseCommand .= ' --verbose';
+        $expectedWoodhouseCommand .= ' --coverage-image artifacts/images/coverage.png';
+        $expectedWoodhouseCommand .= ' --coverage-phpunit artifacts/tests/coverage/coverage.txt';
+        $expectedWoodhouseCommand .= ' --image-theme buckler/buckler';
 
         Phake::when($this->coverallsClient)
             ->exists('Vendor', 'package')
@@ -290,28 +280,18 @@ class BuildCommandTest extends PHPUnit_Framework_TestCase
         $expectedDocumentationCommand = '/path/to/archer/bin/archer documentation';
 
         $expectedCoverallsCommand = '/path/to/project/vendor/bin/coveralls --config';
-        $expectedCoverallsCommand .= " '/path/to/coveralls.yml'";
+        $expectedCoverallsCommand .= " '/path/to/project/.coveralls.yml'";
 
         $expectedWoodhouseCommand  = "/path/to/archer/bin/woodhouse publish 'Vendor/package'";
         $expectedWoodhouseCommand .= ' /path/to/project/artifacts:artifacts';
         $expectedWoodhouseCommand .= ' --message "Publishing artifacts from build 543."';
-        $expectedWoodhouseCommand .= ' --coverage-image artifacts/images/coverage.png';
-        $expectedWoodhouseCommand .= ' --coverage-phpunit artifacts/tests/coverage/coverage.txt';
-        $expectedWoodhouseCommand .= ' --build-status-image artifacts/images/build-status.png';
-        $expectedWoodhouseCommand .= ' --build-status-tap artifacts/tests/report.tap';
         $expectedWoodhouseCommand .= ' --auth-token-env ARCHER_TOKEN';
-        $expectedWoodhouseCommand .= ' --image-theme travis/variable-width';
-        $expectedWoodhouseCommand .= ' --image-theme icecave/regular';
         $expectedWoodhouseCommand .= ' --no-interaction';
         $expectedWoodhouseCommand .= ' --verbose';
 
         Phake::when($this->coverallsClient)
             ->exists('Vendor', 'package')
             ->thenReturn(true);
-
-        Phake::when($this->coverallsConfigManager)
-            ->createConfig(Phake::anyParameters())
-            ->thenReturn('/path/to/coveralls.yml');
 
         Phake::when($this->isolator)
             ->getenv('TRAVIS_PHP_VERSION')
@@ -355,9 +335,11 @@ class BuildCommandTest extends PHPUnit_Framework_TestCase
             Phake::verify($this->output)->writeln('enabled.'),
             Phake::verify($this->isolator)->passthru($expectedTestCommand, 255),
             Phake::verify($this->output)->write('Publishing Coveralls data... '),
+            Phake::verify($this->isolator)->copy('/path/to/archer/res/coveralls/coveralls.yml', '/path/to/project/.coveralls.yml'),
             Phake::verify($this->isolator)->passthru($expectedCoverallsCommand, 255),
             Phake::verify($this->output)->writeln('done.'),
             Phake::verify($this->isolator)->passthru($expectedDocumentationCommand, 255),
+            Phake::verify($this->fileSystem)->delete('/path/to/project/artifacts/tests'),
             Phake::verify($this->isolator)->passthru($expectedWoodhouseCommand, 255)
         );
 
@@ -372,15 +354,12 @@ class BuildCommandTest extends PHPUnit_Framework_TestCase
         $expectedWoodhouseCommand  = "/path/to/archer/bin/woodhouse publish 'Vendor/package'";
         $expectedWoodhouseCommand .= ' /path/to/project/artifacts:artifacts';
         $expectedWoodhouseCommand .= ' --message "Publishing artifacts from build 543."';
-        $expectedWoodhouseCommand .= ' --coverage-image artifacts/images/coverage.png';
-        $expectedWoodhouseCommand .= ' --coverage-phpunit artifacts/tests/coverage/coverage.txt';
-        $expectedWoodhouseCommand .= ' --build-status-image artifacts/images/build-status.png';
-        $expectedWoodhouseCommand .= ' --build-status-tap artifacts/tests/report.tap';
         $expectedWoodhouseCommand .= ' --auth-token-env ARCHER_TOKEN';
-        $expectedWoodhouseCommand .= ' --image-theme travis/variable-width';
-        $expectedWoodhouseCommand .= ' --image-theme icecave/regular';
         $expectedWoodhouseCommand .= ' --no-interaction';
         $expectedWoodhouseCommand .= ' --verbose';
+        $expectedWoodhouseCommand .= ' --coverage-image artifacts/images/coverage.png';
+        $expectedWoodhouseCommand .= ' --coverage-phpunit artifacts/tests/coverage/coverage.txt';
+        $expectedWoodhouseCommand .= ' --image-theme buckler/buckler';
 
         Phake::when($this->coverallsClient)
             ->exists('Vendor', 'package')
@@ -435,15 +414,12 @@ class BuildCommandTest extends PHPUnit_Framework_TestCase
         $expectedWoodhouseCommand  = "/path/to/archer/bin/woodhouse publish 'Vendor/package'";
         $expectedWoodhouseCommand .= ' /path/to/project/artifacts:artifacts';
         $expectedWoodhouseCommand .= ' --message "Publishing artifacts from build 543."';
-        $expectedWoodhouseCommand .= ' --coverage-image artifacts/images/coverage.png';
-        $expectedWoodhouseCommand .= ' --coverage-phpunit artifacts/tests/coverage/coverage.txt';
-        $expectedWoodhouseCommand .= ' --build-status-image artifacts/images/build-status.png';
-        $expectedWoodhouseCommand .= ' --build-status-tap artifacts/tests/report.tap';
         $expectedWoodhouseCommand .= ' --auth-token-env ARCHER_TOKEN';
-        $expectedWoodhouseCommand .= ' --image-theme travis/variable-width';
-        $expectedWoodhouseCommand .= ' --image-theme icecave/regular';
         $expectedWoodhouseCommand .= ' --no-interaction';
         $expectedWoodhouseCommand .= ' --verbose';
+        $expectedWoodhouseCommand .= ' --coverage-image artifacts/images/coverage.png';
+        $expectedWoodhouseCommand .= ' --coverage-phpunit artifacts/tests/coverage/coverage.txt';
+        $expectedWoodhouseCommand .= ' --image-theme buckler/buckler';
 
         Phake::when($this->coverallsClient)
             ->exists('Vendor', 'package')
@@ -496,28 +472,18 @@ class BuildCommandTest extends PHPUnit_Framework_TestCase
         $expectedDocumentationCommand = '/path/to/archer/bin/archer documentation';
 
         $expectedCoverallsCommand = '/path/to/project/vendor/bin/coveralls --config';
-        $expectedCoverallsCommand .= " '/path/to/coveralls.yml'";
+        $expectedCoverallsCommand .= " '/path/to/project/.coveralls.yml'";
 
         $expectedWoodhouseCommand  = "/path/to/archer/bin/woodhouse publish 'Vendor/package'";
         $expectedWoodhouseCommand .= ' /path/to/project/artifacts:artifacts';
         $expectedWoodhouseCommand .= ' --message "Publishing artifacts from build 543."';
-        $expectedWoodhouseCommand .= ' --coverage-image artifacts/images/coverage.png';
-        $expectedWoodhouseCommand .= ' --coverage-phpunit artifacts/tests/coverage/coverage.txt';
-        $expectedWoodhouseCommand .= ' --build-status-image artifacts/images/build-status.png';
-        $expectedWoodhouseCommand .= ' --build-status-tap artifacts/tests/report.tap';
         $expectedWoodhouseCommand .= ' --auth-token-env ARCHER_TOKEN';
-        $expectedWoodhouseCommand .= ' --image-theme travis/variable-width';
-        $expectedWoodhouseCommand .= ' --image-theme icecave/regular';
         $expectedWoodhouseCommand .= ' --no-interaction';
         $expectedWoodhouseCommand .= ' --verbose';
 
         Phake::when($this->coverallsClient)
             ->exists('Vendor', 'package')
             ->thenReturn(true);
-
-        Phake::when($this->coverallsConfigManager)
-            ->createConfig(Phake::anyParameters())
-            ->thenReturn('/path/to/coveralls.yml');
 
         Phake::when($this->isolator)
             ->getenv('TRAVIS_PHP_VERSION')
@@ -561,6 +527,7 @@ class BuildCommandTest extends PHPUnit_Framework_TestCase
             Phake::verify($this->output)->writeln('enabled.'),
             Phake::verify($this->isolator)->passthru($expectedTestCommand, 255),
             Phake::verify($this->output)->write('Publishing Coveralls data... '),
+            Phake::verify($this->isolator)->copy('/path/to/archer/res/coveralls/coveralls.yml', '/path/to/project/.coveralls.yml'),
             Phake::verify($this->isolator)->passthru($expectedCoverallsCommand, 255),
             Phake::verify($this->output)->writeln('failed.'),
             Phake::verify($this->isolator)->passthru($expectedDocumentationCommand, 255),
@@ -575,15 +542,11 @@ class BuildCommandTest extends PHPUnit_Framework_TestCase
         $expectedTestCommand = '/path/to/archer/bin/archer coverage';
 
         $expectedCoverallsCommand = '/path/to/project/vendor/bin/coveralls --config';
-        $expectedCoverallsCommand .= " '/path/to/coveralls.yml'";
+        $expectedCoverallsCommand .= " '/path/to/project/.coveralls.yml'";
 
         Phake::when($this->coverallsClient)
             ->exists('Vendor', 'package')
             ->thenReturn(true);
-
-        Phake::when($this->coverallsConfigManager)
-            ->createConfig(Phake::anyParameters())
-            ->thenReturn('/path/to/coveralls.yml');
 
         Phake::when($this->isolator)
             ->getenv('TRAVIS_PHP_VERSION')
@@ -615,6 +578,7 @@ class BuildCommandTest extends PHPUnit_Framework_TestCase
             Phake::verify($this->output)->writeln('enabled.'),
             Phake::verify($this->isolator)->passthru($expectedTestCommand, 255),
             Phake::verify($this->output)->write('Publishing Coveralls data... '),
+            Phake::verify($this->isolator)->copy('/path/to/archer/res/coveralls/coveralls.yml', '/path/to/project/.coveralls.yml'),
             Phake::verify($this->isolator)->passthru($expectedCoverallsCommand, 255),
             Phake::verify($this->output)->writeln('failed.')
         );