+++ test/suite/Icecave/Archer/Console/Helper/HiddenInputHelperTest.php
@@ -44,7 +44,7 @@ class HiddenInputHelperTest extends PHPUnit_Framework_TestCase
         for ($i = 0; $i < 6; $i ++) {
             $expectedHiddenInputPath = dirname($expectedHiddenInputPath);
         }
-        $expectedHiddenInputPath .= '/lib/Icecave/Archer/Console/Helper/../../../../../res/bin/hiddeninput.exe';
+        $expectedHiddenInputPath .= '/src/Icecave/Archer/Console/Helper/../../../../../res/bin/hiddeninput.exe';
 
         $this->assertSame($expectedHiddenInputPath, $this->_helper->hiddenInputPath());
     }